requirements.txt
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
Django==5.2.10
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
idna==3.11
numpy==2.4.1
pandas==3.0.0
psycopg2-binary==2.9.11
pyjwt==2.12.1
python-dateutil==2.9.0.post0
requests==2.32.5
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
